backend/requirements.txt
asgiref==3.4.1
requests==2.26.0
Django==3.2.14
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt
PyJWT==2.1.0
drf-base64==2.0
fpdf==1.7.2
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.2
pytz==2021.3
reportlab==3.6.3
sqlparse==0.4.2
python-dotenv==0.20.0
djoser==2.1.0
isort==5.10.1
drf-yasg==1.20.0
python-dotenv
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-colorfield==0.7.1